+++ src/util/rng.rs
@@ -12,7 +12,7 @@ impl Dice {
 		}
 	}
 
-	fn roll(&mut self, num_dice: u8, num_sides: u8) -> u8 {
+	fn roll(&mut self, num_dice: i32, num_sides: i32) -> i32 {
 		let mut total = 0;
 		for _ in 0..num_dice {
 			total += self.rng.gen_range(1..num_sides + 1);
@@ -20,15 +20,15 @@ impl Dice {
 		total
 	}
 
-	pub fn roll_3d6(&mut self) -> u8 {
+	pub fn roll_3d6(&mut self) -> i32 {
 		self.roll(3, 6)
 	}
 
-	pub fn roll_1d20(&mut self) -> u8 {
+	pub fn roll_1d20(&mut self) -> i32 {
 		self.roll(1, 20)
 	}
 
-	pub fn from_int(&mut self, num: u32) -> u8 {
-        return self.roll(1, num as u8);
+	pub fn from_int(&mut self, num: u32) -> i32 {
+        return self.roll(1, num as i32);
     }
 }
